requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.0
pytz==2021.3
sqlparse==0.4.2
gunicorn==20.0.4
django-heroku==0.3.1
whitenoise==5.3.0
